fix: TAP-11982 fix for excel load schema npe

+++ connectors-common/file-connector-core/src/main/java/io/tapdata/common/FileConnector.java
@@ -74,7 +74,9 @@ public abstract class FileConnector extends ConnectorBase {
             }
             fileRecordWriter.releaseResource();
         }
-        storage.destroy();
+        if (EmptyKit.isNotNull(storage)) {
+            storage.destroy();
+        }
     }
 
     protected ConcurrentMap<String, TapFile> getFilteredFiles() throws Exception {
